Extract withTest
